moving previousButton and Review Button functionality to ApplicationBaseView
